export the query keys
In this change, I create the query file that export the query keys outside the React function
