+++ pages/comments.tsx
@@ -1,13 +1,6 @@
 import React from "react";
-import { useQuery } from "@tanstack/react-query";
+import { useComments } from "./query"; // Adjust the import path accordingly
 import { inter } from "../styles/fonts";
-
-const getComments = async () => {
-  const COMMENTS_URL = "https://jsonplaceholder.typicode.com/comments";
-
-  const res = await fetch(`${COMMENTS_URL}`);
-  return res.json();
-};
 
 type Comment = {
   id: number;
@@ -16,10 +9,7 @@ type Comment = {
   email: string;
 };
 function Comments() {
-  const { data, error, isLoading } = useQuery({
-    queryKey: ["comments"],
-    queryFn: getComments,
-  });
+  const { data, error, isLoading } = useComments();
 
   if (isLoading) {
     return (
